Refactor details-card component location and imports
Moved details-card.tsx to product-details/details-card.tsx and updated related imports in page.tsx. Also cleaned up formatting in featured-product-list.tsx for better readability.
